# 🐛 bug: Fix MIME type equality checks