--- a/helpers.go
+++ b/helpers.go
@@ -247,8 +247,10 @@ func acceptsOfferType(spec, offerType string, specParams headerParams) bool {
 	s := strings.IndexByte(mimetype, '/')
 	specSlash := strings.IndexByte(spec, '/')
 	// Accept: <MIME_type>/*
-	if specSlash != -1 && utils.EqualFold(spec[:specSlash], mimetype[:s]) && (spec[specSlash:] == "/*" || mimetype[s:] == "/*") {
-		return paramsMatch(specParams, offerParams)
+	if s != -1 && specSlash != -1 {  
+		if utils.EqualFold(spec[:specSlash], mimetype[:s]) && (spec[specSlash:] == "/*" || mimetype[s:] == "/*") {  
+			return paramsMatch(specParams, offerParams)  
+		}
 	}
 
 	return false