--- a/helpers.go
+++ b/helpers.go
@@ -247,9 +247,9 @@ func acceptsOfferType(spec, offerType string, specParams headerParams) bool {
 	s := strings.IndexByte(mimetype, '/')
 	specSlash := strings.IndexByte(spec, '/')
 	// Accept: <MIME_type>/*
-	if s != -1 && specSlash != -1 {  
-		if utils.EqualFold(spec[:specSlash], mimetype[:s]) && (spec[specSlash:] == "/*" || mimetype[s:] == "/*") {  
-			return paramsMatch(specParams, offerParams)  
+	if s != -1 && specSlash != -1 {
+		if utils.EqualFold(spec[:specSlash], mimetype[:s]) && (spec[specSlash:] == "/*" || mimetype[s:] == "/*") {
+			return paramsMatch(specParams, offerParams)
 		}
 	}