## Summary
- check major MIME type using equality in `acceptsOfferType`
- test mismatch when subtype shares prefix