@gaby 2025-07-20
PR:

```
Benchmark_Utils_GetOffer/simple-4                                                   	22355008	        53.61 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:24.5321935Z Benchmark_Utils_GetOffer/6_offers-4                                                 	 8819581	       136.0 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:25.7447605Z Benchmark_Utils_GetOffer/1_parameter-4                                              	 5589475	       216.7 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:26.9757445Z Benchmark_Utils_GetOffer/2_parameters-4                                             	 4150124	       296.3 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:28.1762153Z Benchmark_Utils_GetOffer/3_parameters-4                                             	 3056431	       392.4 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:29.3773126Z Benchmark_Utils_GetOffer/10_parameters-4                                            	  956887	      1254 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:30.5789366Z Benchmark_Utils_GetOffer/6_offers_w/params-4                                        	 2727864	       440.1 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:31.7789703Z Benchmark_Utils_GetOffer/mime_extension-4                                           	 3970009	       302.0 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:05:32.9819136Z Benchmark_Utils_GetOffer/mime_extension#01-4                                        	 2983024	       402.8 ns/op	      48 B/op	       2 allocs/op
2025-07-20T14:05:34.1809424Z Benchmark_Utils_GetOffer/mime_extension#02-4                                        	 2006418	       597.1 ns/op	      48 B/op	       2 allocs/op
2025-07-20T14:05:35.3806766Z Benchmark_Utils_GetOffer/mime_extension#03-4                                        	 4346067	       275.9 ns/op	       0 B/op	       0 allocs/op
```

main

```
Benchmark_Utils_GetOffer/simple-4                                                   	22340031	        53.60 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:21.3740715Z Benchmark_Utils_GetOffer/6_offers-4                                                 	 9218581	       129.8 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:22.5788930Z Benchmark_Utils_GetOffer/1_parameter-4                                              	 5554348	       216.7 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:23.7857590Z Benchmark_Utils_GetOffer/2_parameters-4                                             	 4089624	       294.8 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:24.9914949Z Benchmark_Utils_GetOffer/3_parameters-4                                             	 3085041	       390.5 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:26.1878443Z Benchmark_Utils_GetOffer/10_parameters-4                                            	  949168	      1259 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:27.3907517Z Benchmark_Utils_GetOffer/6_offers_w/params-4                                        	 2422017	       496.1 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:28.5880289Z Benchmark_Utils_GetOffer/mime_extension-4                                           	 3815960	       313.5 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:29.7873809Z Benchmark_Utils_GetOffer/mime_extension#01-4                                        	 3068139	       390.6 ns/op	      48 B/op	       2 allocs/op
2025-07-20T14:14:30.9868465Z Benchmark_Utils_GetOffer/mime_extension#02-4                                        	 2025990	       591.6 ns/op	      48 B/op	       2 allocs/op
2025-07-20T14:14:32.1863975Z Benchmark_Utils_GetOffer/mime_extension#03-4                                        	 4360053	       274.9 ns/op	       0 B/op	       0 allocs/op
2025-07-20T14:14:33.3854392Z Benchmark_Utils_GetOffer/web_browser-4                                              	 5981464	       200.3 ns/op	       0 B/op	       0 allocs/op
```